replace model/jsonschema/*.kt with bean/jsonschema/*.kt
